fix: emulate connectedCallback on construct
