+++ src/Component.ts
@@ -1,8 +1,8 @@
 import { document, HTMLElement } from './window';
 import { ComponentInterface, ComponentConstructorInterface, COMPONENT_SYMBOL } from './Interfaces';
 import { customElements } from './CustomElementRegistry';
-import { isElement, removeChildImpl, setAttributeImpl, nextTick } from './helpers';
-import { DOM, isConnected, connect, emulateLifeCycle, shouldEmulateLifeCycle } from './DOM';
+import { isElement, isConnected, emulateLifeCycle, removeChildImpl, setAttributeImpl } from './helpers';
+import { DOM } from './DOM';
 import { DelegatedEventCallback, delegateEventListener, undelegateEventListener, dispatchEvent, dispatchAsyncEvent, getListeners } from './events';
 import { getContext, createContext } from './Context';
 import { Template } from './Template';
@@ -55,11 +55,6 @@ const mixin = <T extends typeof HTMLElement>(constructor: T) => class Component 
     }
 
     /**
-     * Check if the element is still constructing on connectedCallback.
-     */
-    private constructing?: boolean;
-
-    /**
      * Create a new Component instance.
      * @param node Instantiate the element using the given node instead of creating a new one.
      * @param properties A set of initial properties for the element.
@@ -110,15 +105,6 @@ const mixin = <T extends typeof HTMLElement>(constructor: T) => class Component 
             }
         }
 
-        if (element.isConnected) {
-            if (shouldEmulateLifeCycle(element) || element !== this) {
-                this.constructing = true;
-                connect(element, true);
-                this.constructing = false;
-                nextTick(() => this.forceUpdate());
-            }
-        }
-
         return element;
     }
 
@@ -130,9 +116,7 @@ const mixin = <T extends typeof HTMLElement>(constructor: T) => class Component 
         // force the is attribute for styling
         setAttributeImpl.call(this, 'is', this.is);
         // trigger a re-render when the Node is connected
-        if (!this.constructing) {
-            this.forceUpdate();
-        }
+        this.forceUpdate();
     }
 
     /**
