+++ src/helpers.ts
@@ -1,4 +1,8 @@
 import { window, document, CustomEvent, Node, HTMLElement, Event } from './window';
+import { createSymbolKey } from './symbols';
+import { ComponentInterface } from './Interfaces';
+import { cloneChildNodes } from './NodeList';
+
 const { DOCUMENT_NODE, TEXT_NODE, COMMENT_NODE, ELEMENT_NODE } = Node;
 
 /**
@@ -129,3 +133,86 @@ export const isEvent = (event: any): event is Event => event instanceof Event;
  * Run a task at the next tick.
  */
 export const nextTick = window.requestAnimationFrame || setTimeout;
+
+/**
+ * Check if a Node is connected.
+ *
+ * @return A truthy value for connected targets.
+ */
+export const isConnected: (this: Node | null) => boolean = isConnectedImpl ? (isConnectedImpl as any).get : function(this: Node | null): boolean {
+    if (isElement(this) || isText(this)) {
+        return isConnected.call(this.parentNode);
+    }
+    if (isDocument(this)) {
+        return true;
+    }
+
+    return false;
+};
+
+/**
+ * Invoke `connectedCallback` method of a Node (and its descendents).
+ * It does nothing if life cycle is disabled.
+ *
+ * @param node The connected node.
+ */
+export const connect = (node: Node, force = false) => {
+    if (!isElement(node)) {
+        return;
+    }
+    if (force || shouldEmulateLifeCycle(node)) {
+        (node as ComponentInterface<HTMLElement>).connectedCallback();
+    }
+    let children = cloneChildNodes(node.childNodes);
+    for (let i = 0, len = children.length; i < len; i++) {
+        connect(children[i]);
+    }
+};
+
+/**
+ * Invoke `disconnectedCallback` method of a Node (and its descendents).
+ * It does nothing if life cycle is disabled.
+ *
+ * @param node The disconnected node.
+ */
+export const disconnect = (node: Node) => {
+    if (!isElement(node)) {
+        return;
+    }
+    if (shouldEmulateLifeCycle(node)) {
+        node.disconnectedCallback();
+    }
+    let children = cloneChildNodes(node.childNodes);
+    for (let i = 0, len = children.length; i < len; i++) {
+        disconnect(children[i]);
+    }
+};
+
+/**
+ * A symbol which identify emulated components.
+ */
+const EMULATE_LIFECYCLE_SYMBOL = createSymbolKey();
+
+/**
+ * Check if a node require emulated life cycle.
+ * @param node The node to check.
+ */
+export const shouldEmulateLifeCycle = (node: any): node is ComponentInterface<HTMLElement> => node[EMULATE_LIFECYCLE_SYMBOL];
+
+/**
+ * Should emulate life cycle.
+ */
+let lifeCycleEmulation = typeof customElements === 'undefined';
+
+/**
+ * Flag the element for life cycle emulation.
+ */
+export const emulateLifeCycle = (node: HTMLElement) => {
+    lifeCycleEmulation = true;
+    (node as any)[EMULATE_LIFECYCLE_SYMBOL] = true;
+};
+
+/**
+ * Life cycle emulation status.
+ */
+export const emulatingLifeCycle = () => lifeCycleEmulation;
