+++ src/index.ts
@@ -3,7 +3,8 @@ export { window };
 export { customElements, CustomElementRegistry } from './CustomElementRegistry';
 export { Observable } from './Observable';
 export { TemplateItem, TemplateItems, Template, TemplateFilter, TemplateFunction } from './Template';
-export { DOM, connect, disconnect } from './DOM';
+export { connect, disconnect } from './helpers';
+export { DOM } from './DOM';
 export { Context } from './Context';
 export { render } from './render';
 export { Fragment } from './Fragment';
